+++ src/lib.rs
@@ -41,6 +41,7 @@ mod raw {
 
 //
 pub mod devices;
+pub mod error;
 pub mod extensions;
 pub mod modules;
 pub mod types;